+++ test/utils/createApiResponse.ts
@@ -63,6 +63,3 @@ export function createSFFVResponse(
     ...args,
   };
 }
-
-export const forceDelay = (ms: number) =>
-  new Promise((resolve) => setTimeout(resolve, ms));
