+++ test/utils/fixtures.ts
@@ -1,0 +1,3 @@
+export const forceDelay = (ms: number) =>
+  new Promise((resolve) => setTimeout(resolve, ms));
+
